Add a lot of platform stuff (#274)

+++ scripts/common.sh
@@ -145,7 +145,10 @@ on_chroot() {
         mount --bind /etc/resolv.conf "${MNT_DIR}/etc/resolv.conf"
     fi
 
+    if [ ! -e ${MNT_DIR}/opt/additionalFiles ]; then
+    df -h
     cp -r "${STAGE_DIR}/../../additionalFiles" "${MNT_DIR}/opt"
+    fi
     #sudo chroot --userspec=1000:1000 "$MNT_DIR" /bin/bash "/home/pi/install.sh"
     capsh --drop=cap_setfcap "--chroot=${MNT_DIR}/" -- "$@"
